Crate fedimint_recurringd

Modules

db 🔒

Structs

InvoiceStatus

Helper struct indicating if an invoice was paid. In the future it may also contain the preimage to be fully LUD-21 compliant.

LNURLPayInvoice

The lnurl-rs crate doesn’t have the verify field in this type and we don’t use any of the other fields right now. Once we upstream the verify field this struct can be removed.

RecurringInvoiceServer

Enums

PaymentCodeInvoice

PaymentStatus

Traits

LnClientContextExt 🔒

Functions

await_invoice_confirmed 🔒

operation_id_from_user_key 🔒